Internalize yaml config and swapper
