+++ wormhole-console/src/main/java/com/zergclan/wormhole/console/infra/config/WormholeConfigurationLoader.java
@@ -19,9 +19,9 @@ package com.zergclan.wormhole.console.infra.config;
 
 import com.zergclan.wormhole.common.constant.MarkConstant;
 import com.zergclan.wormhole.config.core.WormholeConfiguration;
-import com.zergclan.wormhole.console.infra.config.swapper.YamlWormholeConfigurationSwapper;
-import com.zergclan.wormhole.console.infra.config.yaml.YamlTaskConfiguration;
-import com.zergclan.wormhole.console.infra.config.yaml.YamlWormholeConfiguration;
+import com.zergclan.wormhole.config.core.swapper.YamlWormholeConfigurationSwapper;
+import com.zergclan.wormhole.config.core.yaml.YamlTaskConfiguration;
+import com.zergclan.wormhole.config.core.yaml.YamlWormholeConfiguration;
 import com.zergclan.wormhole.console.infra.util.YamlUtil;
 import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
