+++ wormhole-console/src/main/resources/conf/wormhole-task-task_two.yaml
@@ -31,6 +31,7 @@ source:
 target:
   dataSource: ds-target
   table: t_user
+  transaction: false
   uniqueNodes: id,name
   compareNodes: id,name
   ignoreNodes: create_time,update_time
