+++ wormhole-infra/wormhole-config/src/main/java/com/zergclan/wormhole/config/core/yaml/YamlDataNodeConfiguration.java
@@ -15,23 +15,22 @@
  * limitations under the License.
  */
 
-package com.zergclan.wormhole.console.infra.config.yaml;
+package com.zergclan.wormhole.config.core.yaml;
 
+import com.zergclan.wormhole.config.api.YamlConfiguration;
 import lombok.Getter;
 import lombok.Setter;
 
-import java.util.Properties;
-
 /**
- * YAML filter configuration.
+ * YAML data source configuration.
  */
 @Getter
 @Setter
-public final class YamlFilterConfiguration implements YamlConfiguration {
+public final class YamlDataNodeConfiguration implements YamlConfiguration {
     
-    private String type;
+    private String nodeType;
     
-    private int order;
+    private String dataType;
     
-    private Properties prop;
+    private String defaultValue;
 }
