+++ wormhole-infra/wormhole-config/src/main/java/com/zergclan/wormhole/config/core/yaml/YamlTargetConfiguration.java
@@ -15,24 +15,33 @@
  * limitations under the License.
  */
 
-package com.zergclan.wormhole.console.infra.config.yaml;
+package com.zergclan.wormhole.config.core.yaml;
 
+import com.zergclan.wormhole.config.api.YamlConfiguration;
 import lombok.Getter;
 import lombok.Setter;
 
-import java.util.Collection;
-import java.util.LinkedList;
+import java.util.LinkedHashMap;
+import java.util.Map;
 
 /**
- * YAML data node mapping configuration.
+ * YAML target configuration.
  */
 @Getter
 @Setter
-public final class YamlDataNodeMappingConfiguration implements YamlConfiguration {
+public final class YamlTargetConfiguration implements YamlConfiguration {
     
-    private String targetNames;
+    private String dataSource;
     
-    private String sourceNames;
+    private String table;
     
-    private Collection<YamlFilterConfiguration> filters = new LinkedList<>();
+    private boolean transaction;
+    
+    private String uniqueNodes;
+    
+    private String compareNodes;
+    
+    private String ignoreNodes;
+    
+    private Map<String, YamlDataNodeConfiguration> dataNodes = new LinkedHashMap<>();
 }
