+++ wormhole-infra/wormhole-config/src/main/java/com/zergclan/wormhole/config/core/yaml/YamlTaskConfiguration.java
@@ -15,28 +15,29 @@
  * limitations under the License.
  */
 
-package com.zergclan.wormhole.console.infra.config.yaml;
+package com.zergclan.wormhole.config.core.yaml;
 
+import com.zergclan.wormhole.config.api.YamlConfiguration;
 import lombok.Getter;
 import lombok.Setter;
 
-import java.util.LinkedHashMap;
-import java.util.Map;
+import java.util.Collection;
+import java.util.LinkedList;
 
 /**
- * YAML source configuration.
+ * YAML task configuration.
  */
 @Getter
 @Setter
-public final class YamlSourceConfiguration implements YamlConfiguration {
+public final class YamlTaskConfiguration implements YamlConfiguration {
     
-    private String dataSource;
+    private int order;
     
-    private String actualSql;
+    private int batchSize;
     
-    private String table;
+    private YamlSourceConfiguration source;
     
-    private String conditionSql;
+    private YamlTargetConfiguration target;
     
-    private Map<String, YamlDataNodeConfiguration> dataNodes = new LinkedHashMap<>();
+    private Collection<YamlDataNodeMappingConfiguration> dataNodeMappings = new LinkedList<>();
 }
